Improve test coverage for QDC records.
Consolidates QDC tests into a single class to make it easier to ensure code coverage.
