+++ module/Finna/tests/fixtures/qdc/qdc_museum_test_filtered.xml
@@ -7,7 +7,6 @@
   <description lang="fi" comment="Wikidata description">maalaus: Kuoma, Juha</description>
   <description lang="sv" comment="Wikidata description">m&#xE5;lning av Juha Kuoma</description>
   <description lang="en" comment="Wikidata description">painting by Juha Kuoma</description>
-  
   <creator href="http://www.wikidata.org/entity/Q11868768" comment="">Juha Kuoma</creator>
   <publisher href="http://www.wikidata.org/entity/Q4412746" comment="P276 (location)">Savannin taidepaja</publisher>
   <date comment="P571 (inception)&#13;&#10;">2022</date>
